# PERF: avoid zeros_like in groupby.pyx