Apparently calling `np.zeros_like` has quite some overhead:

```
In [7]: arr = np.random.randn(1000)

In [8]: %timeit np.zeros_like(arr)
3.14 µs ± 44.2 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [9]: %timeit np.zeros(arr.shape, dtype=arr.dtype)
662 ns ± 41.9 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
``` 

So this PR replaces a few (recently added) occurences in groupby.pyx.

Using the same benchmarks case as in https://github.com/pandas-dev/pandas/pull/40178#issue-583301313, this gives:

```
In [2]: %timeit df_am.groupby(labels).sum()
66.5 ms ± 876 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)   <--- master
54.6 ms ± 958 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)   <--- PR
```